backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
requests==2.31.0
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
